+++ src/mlrb/MyLittleRangeBook/MyLittleRangeBook.csproj
@@ -31,8 +31,4 @@
         <EmbeddedResource Include="Persistence\Sqlite\Scripts\*.sql"/>
     </ItemGroup>
 
-    <ItemGroup>
-        <Folder Include="Console\"/>
-    </ItemGroup>
-
 </Project>
